requirements.txt
beautifulsoup4==4.6.0
bleach==2.0.0
DateTime==4.0.1
defusedxml
-e git+https://github.com/bbusenius/Diablo-Python.git#egg=diablo_python 
Django==2.0.3
django-appconf==1.0.2
django-bleach==0.3.0
django-compressor==2.1
django-cors-headers==2.2.0
django-libsass==0.6
django-localflavor==1.2
django-macros==0.4.0
django-modelcluster==4.1
django-redis
-e git+https://github.com/Brown-University-Library/django-shibboleth-remoteuser.git#egg=django_shibboleth_remoteuser
django-static-precompiler==1.8.2
django-taggit==0.22.2
django-treebeard==4.2.1
djangorestframework==3.7.7
elasticsearch>=5.0.0,<6.0.0
feedparser==5.2.1
html5lib==0.999999999
libsass==0.10.1
lxml==3.5.0
networkx
openpyxl
Pillow==4.3.0
psycopg2==2.7.1
Pygments==2.0.2
pytz==2018.3
rcssmin==1.0.6
requests==2.18.4
rjsmin==1.0.12
simplejson==3.11.1
six==1.11.0
sqlparse==0.2.4
Unidecode==0.4.21
urllib3==1.22
wagtail==2.0
-e git+https://github.com/Rotund/wagtailmedia.git@e687233343c1e7530e73a008e68ac854ee56b5ed#egg=wagtailmedia
wheel==0.24.0
Willow==1.1
zope.interface==4.1.2
